ssh方式先删除再clone

+++ application/src/main/java/com/dr/code/diff/util/GitRepoUtil.java
@@ -130,7 +130,10 @@ public class GitRepoUtil {
         Git git = null;
         try {
             //ssh无法读取本地代码仓，先删除再clone
-            FileUtil.removeDir(new File(codePath));
+            File repoGitDir = new File(codePath + GitUrlTypeEnum.SSH.getValue() + "/.git");
+            if (repoGitDir.exists()) {
+                FileUtil.removeDir(new File(codePath));
+            }
             LoggerUtil.info(log, "本地代码不存在，clone", gitUrl, codePath);
             //为了区分ssh路径和http/s这里ssh多加了一层目录
             git = instanceSshGit(gitUrl, codePath + GitUrlTypeEnum.SSH.getValue(), commitId, privateKey);
